splitted database operations into different files
